implementing best practices in the controller

+++ Controllers/ProdutosController.cs
@@ -4,6 +4,7 @@ using apifuncional.Data;
 using apifuncional.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Http;
 
 namespace apifuncional.Controllers;
 
@@ -22,22 +23,51 @@ public class ProdutosController : ControllerBase
     }
 
     [HttpGet]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesDefaultResponseType]
     public async Task<ActionResult<IEnumerable<Produto>>> GetProdutos()
     {
+        if (_context.Produtos == null)
+        {
+            return NotFound();
+        }
+
         return await _context.Produtos.ToListAsync();
     }
 
     [HttpGet("{id:int}")]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesDefaultResponseType]
     public async Task<ActionResult<Produto>> GetProduto(int id)
     {
+        if (_context.Produtos == null)
+        {
+            return NotFound();
+        }
+
         var produto = await _context.Produtos.FindAsync(id);
+
+        if (produto == null)
+        {
+            return NotFound();
+        }
 
         return produto;
     }
 
     [HttpPost]
+    [ProducesResponseType(StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesDefaultResponseType]
     public async Task<ActionResult<Produto>> PostProduto(Produto produto)
     {
+        if (_context.Produtos == null)
+        {
+            return Problem("Erro ao cadastrar um produto!");
+        }
+
         if (!ModelState.IsValid)
         {
             // return BadRequest(ModelState);
@@ -58,6 +88,10 @@ public class ProdutosController : ControllerBase
     }
 
     [HttpPut("{id:int}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesDefaultResponseType]
     public async Task<IActionResult> PutProduto(int id, Produto produto)
     {
 
@@ -65,21 +99,54 @@ public class ProdutosController : ControllerBase
 
         if (!ModelState.IsValid) return ValidationProblem(ModelState);
 
-        _context.Produtos.Update(produto);
-        await _context.SaveChangesAsync();
+        _context.Entry(produto).State = EntityState.Modified;
+
+        try
+        {
+            await _context.SaveChangesAsync();
+        }
+        catch (DbUpdateConcurrencyException)
+        {
+            if (!ProdutoExists(id))
+            {
+                return NotFound();
+            }
+            else
+            {
+                throw;
+            }
+        }
 
         return NoContent();
     }
 
     [HttpDelete("{id:int}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesDefaultResponseType]
     public async Task<IActionResult> DeleteProduto(int id)
     {
+        if (_context.Produtos == null)
+        {
+            return NotFound();
+        }
+
         var produto = await _context.Produtos.FindAsync(id);
+
+        if (produto == null)
+        {
+            return NotFound();
+        }
 
         _context.Produtos.Remove(produto);
         await _context.SaveChangesAsync();
 
         return NoContent();
+    }
+
+    public bool ProdutoExists(int id)
+    {
+        return (_context.Produtos?.Any(e => e.Id == id)).GetValueOrDefault();
     }
